src/requirements.txt
pytest
pytest-mock
ruff
dbt-core==1.8.4
dbt-postgres
astronomer-cosmos==1.11.0
